+++ routes/login.js
@@ -1,7 +1,8 @@
-const express = require('express');
+const express = require("express");
 const router = express.Router();
-const { upload, uploadFileToGCS } = require('../utils/UploadFile');
+const { upload } = require("../utils/UploadFile");
+const { uploadFileToGCS } = require("../controllers/loginControllers");
 
-router.post('/upload', upload.single('file'), uploadFileToGCS);
+router.post("/upload", upload.single("file"), uploadFileToGCS);
 
 module.exports = router;
