+++ utils/UploadFile.js
@@ -24,39 +24,7 @@ const storageClient = new Storage({
 const bucketName = process.env.BUCKET_NAME;
 const bucket = storageClient.bucket(bucketName);
 
-const uploadFileToGCS = async (req, res) => {
-  if (!req.file) {
-    return res.status(400).send("No file uploaded.");
-  }
-
-  const blob = bucket.file(req.file.originalname);
-  const blobStream = blob.createWriteStream({
-    resumable: false,
-  });
-
-  blobStream.on("error", (err) => {
-    console.error(err);
-    res.status(500).send("Error uploading file");
-  });
-
-  blobStream.on("finish", async () => {
-    try {
-      const [url] = await blob.getSignedUrl({
-        action: "read",
-        expires: "01-01-2100",
-      });
-
-      res.status(200).send(`File uploaded successfully. Public URL: ${url}`);
-    } catch (err) {
-      console.error(err);
-      res.status(500).send("Error generating signed URL");
-    }
-  });
-
-  blobStream.end(req.file.buffer);
-};
-
 module.exports = {
   upload,
-  uploadFileToGCS,
+  bucket,
 };
